perf: 优化文件管理的冗余代码
