+++ pkg/jms-sdk-go/model/node_tree.go
@@ -1,26 +1,56 @@
 package model
 
-import "encoding/json"
+import "strings"
 
-type NodeTreeList []NodeTreeAsset
+type NodeTreeList []NodeTree
 
-type NodeTreeAsset struct {
-	ID       string                 `json:"id"`
-	Name     string                 `json:"name"`
-	Title    string                 `json:"title"`
-	Pid      string                 `json:"pId"`
-	IsParent bool                   `json:"isParent"`
-	Meta     map[string]interface{} `json:"meta"`
+type NodeTree struct {
+	ID       string   `json:"id"`
+	Name     string   `json:"name"`
+	Title    string   `json:"title"`
+	Pid      string   `json:"pId"`
+	IsParent bool     `json:"isParent"`
+	Meta     TreeMeta `json:"meta"`
 
 	ChkDisabled bool `json:"chkDisabled"` // 判断资产是否禁用
 }
 
-func ConvertMetaToNode(body []byte) (node Node, err error) {
-	err = json.Unmarshal(body, &node)
-	return
+type TreeMeta struct {
+	Type string       `json:"type"`
+	Data NodeTreeMeta `json:"data"`
 }
 
-func ConvertMetaToAsset(body []byte) (asset Asset, err error) {
-	err = json.Unmarshal(body, &asset)
-	return
+type NodeTreeMeta struct {
+	ID string `json:"id"`
+
+	NodeMeta
+	AssetMeta
 }
+
+func (n NodeTreeMeta) IsSupportProtocol(protocol string) bool {
+	for _, item := range n.Protocols {
+		if strings.Contains(strings.ToLower(item),
+			strings.ToLower(protocol)) {
+			return true
+		}
+	}
+	return false
+}
+
+type NodeMeta struct {
+	Key   string `json:"key"`
+	Value string `json:"value"`
+}
+
+type AssetMeta struct {
+	Hostname  string   `json:"hostname"`
+	IP        string   `json:"ip"`
+	Protocols []string `json:"protocols"`
+	Platform  string   `json:"platform"`
+	OrgName   string   `json:"org_name"`
+}
+
+const (
+	TreeTypeNode  = "node"
+	TreeTypeAsset = "asset"
+)
