+++ pkg/utils/stat_test.go
@@ -5,7 +5,7 @@ import (
 )
 
 func TestBasic(t *testing.T) {
-	t.Log("load cpu==> ",CpuLoad1Usage())
+	t.Log("load cpu==> ", CpuLoad1Usage())
 	t.Log("men==> ", MemoryUsagePercent())
 	t.Log("disk==> ", DiskUsagePercent())
 	t.Log("local ip ==> ", CurrentLocalIP())
